Feat [#27] : Redis cluster 관련 config 파일 작성
