+++ src/main/java/com/sparta/ezpzuser/common/config/redis/RedisClusterProperties.java
@@ -1,0 +1,20 @@
+package com.sparta.ezpzuser.common.config.redis;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+import java.util.List;
+
+@Data
+@Component
+@NoArgsConstructor
+@ConfigurationProperties(prefix = "spring.data.redis.cluster")
+public class RedisClusterProperties {
+
+    private int maxRedirects;
+    private String connectIp;
+    private List<String> nodes;
+
+}
