--- src/test/java/com/sparta/ezpzuser/EzpzUserApplicationTests.java
@@ -1,12 +1,0 @@
-package com.sparta.ezpzuser;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-public class EzpzUserApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-}
